+++ CHANGES.rst
@@ -12,6 +12,9 @@ Backwards incompatible changes
   version 6.0. If you are still using it, switch to
   `Guided package installation` as described in `Upgrade installation`_.
 
+- Integrate `src/icemac/addressbook/base.zcml` into
+  `src/icemac/addressbook/roles.zcml`.
+
 Features
 --------